requirements.txt
python-inquirer
click
pathlib
pillow